requirements.txt
flask
requests
python-dotenv
pandas
openai
langchain==0.2.14
langchain-openai==0.1.7
